+++ src/options.ts
@@ -1,7 +1,6 @@
-import { ParserOptions, SupportOption } from 'prettier';
+import { ParserOptions as PrettierParserOptions, SupportOption } from 'prettier';
 
-declare module 'prettier' {
-    interface RequiredOptions extends PluginOptions {}
+export interface ParserOptions<T = any> extends PrettierParserOptions<T>, Partial<PluginOptions> {
 }
 
 export interface PluginOptions {
@@ -18,7 +17,6 @@ function makeChoice(choice: string) {
 
 export const options: Record<keyof PluginOptions, SupportOption> = {
     svelteSortOrder: {
-        since: '0.6.0',
         category: 'Svelte',
         type: 'choice',
         default: 'options-scripts-markup-styles',
@@ -59,21 +57,18 @@ export const options: Record<keyof PluginOptions, SupportOption> = {
         ],
     },
     svelteStrictMode: {
-        since: '0.7.0',
         category: 'Svelte',
         type: 'boolean',
         default: false,
         description: 'More strict HTML syntax: self-closed tags, quotes in attributes',
     },
     svelteBracketNewLine: {
-        since: '0.6.0',
         category: 'Svelte',
         type: 'boolean',
         description: 'Put the `>` of a multiline element on a new line',
         deprecated: '2.5.0',
     },
     svelteAllowShorthand: {
-        since: '1.0.0',
         category: 'Svelte',
         type: 'boolean',
         default: true,
@@ -81,7 +76,6 @@ export const options: Record<keyof PluginOptions, SupportOption> = {
             'Option to enable/disable component attribute shorthand if attribute name and expressions are same',
     },
     svelteIndentScriptAndStyle: {
-        since: '1.2.0',
         category: 'Svelte',
         type: 'boolean',
         default: true,
@@ -130,7 +124,7 @@ export type SortOrderPart = 'scripts' | 'markup' | 'styles' | 'options';
 
 const sortOrderSeparator = '-';
 
-export function parseSortOrder(sortOrder: SortOrder): SortOrderPart[] {
+export function parseSortOrder(sortOrder: SortOrder = 'options-scripts-markup-styles'): SortOrderPart[] {
     if (sortOrder === 'none') {
         return [];
     }