+++ test/printer/index.ts
@@ -1,6 +1,7 @@
 import test from 'ava';
 import { readdirSync, readFileSync, existsSync } from 'fs';
 import { format } from 'prettier';
+import * as SveltePlugin from '../../src'
 
 let files = readdirSync('test/printer/samples').filter(
     (name) => name.endsWith('.html') || name.endsWith('.md'),
@@ -17,16 +18,16 @@ for (const file of files) {
     const ending = file.split('.').pop();
     const input = readFileSync(`test/printer/samples/${file}`, 'utf-8').replace(/\r?\n/g, '\n');
     const options = readOptions(
-        `test/printer/samples/${file.replace(`.${ending}`, '.options.json')}`,
+        `test/printer/samples/${file.replace('.only', '').replace(`.${ending}`, '.options.json')}`,
     );
 
-    test(`printer: ${file.slice(0, file.length - `.${ending}`.length)}`, (t) => {
-        const actualOutput = format(input, {
-            parser: (ending === 'html' ? 'svelte' : 'markdown') as any,
-            plugins: [require.resolve('../../src')],
+    test(`printer: ${file.slice(0, file.length - `.${ending}`.length)}`, async (t) => {
+        const actualOutput = await format(input, {
+            parser: (ending === 'html' ? 'svelte' : 'markdown'),
+            plugins: [SveltePlugin],
             tabWidth: 4,
             ...options,
-        } as any);
+        });
 
         t.is(input, actualOutput, `Expected:\n${input}\n\nActual:\n${actualOutput}`);
     });