+++ CHANGELOG.md
@@ -1,9 +1,21 @@
 # prettier-plugin-svelte changelog
 
-## 3.0.0 (Unreleased
+## 3.0.0 (Unreleased)
 
 -   (breaking) requires `prettier` version 3
 -   (breaking) requires node version 14 or higher
+
+## 2.10.1
+
+-   (chore) mark as compatible with Svelte 4
+
+## 2.10.0
+
+-   (feat) support `requirePragma` and `insertPragma` options ([#350](https://github.com/sveltejs/prettier-plugin-svelte/issues/350))
+-   (feat) support `<svelte:document>`
+-   (feat) trim whitespace in `class` attributes ([#339](https://github.com/sveltejs/prettier-plugin-svelte/issues/339))
+-   (feat) allow multiple comments atop of script/style tags ([#291](https://github.com/sveltejs/prettier-plugin-svelte/issues/291))
+-   (fix) handle script/style attributes without quotes ([#344](https://github.com/sveltejs/prettier-plugin-svelte/issues/344))
 
 ## 2.9.0